back/requirements.txt
fastapi
uvicorn
google-generativeai
pandas
openpyxl
tabulate